FoodApp Croma

+++ app/src/main/java/com/croma/app/foodApp/Fooddetail.java
@@ -6,10 +6,10 @@ package com.croma.app.foodApp;
 public class Fooddetail {
     public String foodName;
     public int forwardImg;
-    public double foodPrice;
+    public int foodPrice;
 
 
-    Fooddetail(String foodName, int forwardImg, double foodPrice){
+    Fooddetail(String foodName, int forwardImg, int foodPrice){
         this.foodName=foodName;
         this.forwardImg=forwardImg;
         this.foodPrice=foodPrice;
